+++ tests/primary_query_test.py
@@ -249,6 +249,7 @@ def test_mini_1_1():
     assert len(w.lexicons()) == 1
     assert len(w.expanded_lexicons()) == 1
     assert len(w.synsets('例え')[0].hypernyms()) == 1
+    assert w.synsets('例え')[0].lexfile() == 'noun.cognition'
     assert len(w.words('例え')[0].lemma().pronunciations()) == 1
     p = w.words('例え')[0].lemma().pronunciations()[0]
     assert p.value == 'tatoe'