+++ wn/lmf.py
@@ -29,6 +29,7 @@ from wn.constants import (
     SYNSET_RELATIONS,
     ADJPOSITIONS,
     PARTS_OF_SPEECH,
+    LEXICOGRAPHER_FILES,
 )
 
 
@@ -215,7 +216,8 @@ class Definition(_HasMeta):
 
 class Synset(_HasMeta):
     __slots__ = ('id', 'ili', 'pos', 'definitions', 'ili_definition',
-                 'relations', 'examples', 'lexicalized', 'members', 'external')
+                 'relations', 'examples', 'lexicalized', 'members',
+                 'lexfile', 'external')
 
     def __init__(
         self,
@@ -228,6 +230,7 @@ class Synset(_HasMeta):
         examples: List[Example] = None,
         lexicalized: bool = True,
         members: List[str] = None,
+        lexfile: str = None,
         meta: Metadata = None
     ):
         super().__init__(meta)
@@ -240,6 +243,7 @@ class Synset(_HasMeta):
         self.examples = examples or []
         self.lexicalized = lexicalized
         self.members = members or []
+        self.lexfile = lexfile
         self.external = False
 
     @classmethod
@@ -794,6 +798,7 @@ def _load_synsets(
                 examples=_load_examples(events, version),
                 lexicalized=_get_bool(attrs.get('lexicalized', 'true')),
                 members=attrs.get('members', '').split(),
+                lexfile=_get_literal(attrs.get('lexfile'), LEXICOGRAPHER_FILES),
                 meta=_get_metadata(attrs, version),
             )
             events.end('Synset')
@@ -1108,8 +1113,11 @@ def _dump_synset(synset: Synset, out: TextIO, version: str) -> None:
             attrib['partOfSpeech'] = synset.pos
         if not synset.lexicalized:
             attrib['lexicalized'] = 'false'
-        if version != '1.0' and synset.members:
-            attrib['members'] = ' '.join(synset.members)
+        if version != '1.0':
+            if synset.members:
+                attrib['members'] = ' '.join(synset.members)
+            if synset.lexfile:
+                attrib['lexfile'] = synset.lexfile
         attrib.update(_meta_dict(synset.meta))
         elem = ET.Element('Synset', attrib=attrib)
     elem.extend(_build_definition(defn) for defn in synset.definitions)